requirements.txt
beautifulsoup4
requests
pyTelegramBotAPI
schedule
python-dotenv
